fix: access_token timestamp

+++ backend/app/core/config.py
@@ -15,7 +15,7 @@ class Settings(BaseSettings):
     JWT_AT_KEY: str
     JWT_SECRET_KEY: str
     JWT_ALGORITHM: str
-    JWT_ACCESS_TOKEN_EXPIRE_MINUTES: int
+    JWT_ACCESS_TOKEN_EXPIRE_MINUTES: float
 
     @property
     def SQLALCHEMY_DATABASE_URI(self) -> str:
